+++ testcases/kernel/syscalls/mmap/mmap03.c
@@ -175,7 +175,7 @@ static void setup(void)
 	}
 
 	/* Write test buffer contents into temporary file */
-	if (write(fildes, tst_buff, page_sz) < page_sz) {
+	if (write(fildes, tst_buff, page_sz) < (long)page_sz) {
 		free(tst_buff);
 		tst_brkm(TFAIL | TERRNO, cleanup, "writing to %s failed",
 			 TEMPFILE);
